+++ src/main/java/mm/rest/TeamsReps.java
@@ -11,13 +11,13 @@ public class TeamsReps {
     public String id;
     @SerializedName("create_at")
     @Expose
-    public int createAt;
+    public long createAt;
     @SerializedName("update_at")
     @Expose
-    public int updateAt;
+    public long updateAt;
     @SerializedName("delete_at")
     @Expose
-    public int deleteAt;
+    public long deleteAt;
     @SerializedName("display_name")
     @Expose
     public String displayName;
